url-generator/requirements.txt
Flask==3.1.0
minio==7.2.15
python-dotenv==1.1.0
